Fix .ssh symbolic link duplication bug

+++ root/etc/cont-init.d/20-init-openssh-server.sh
@@ -11,17 +11,21 @@ mkdir -p /mnt/data/ssh/config
 mkdir -p /mnt/data/ssh/.ssh
 
 # Copy default config
-if [ ! -e /mnt/data/ssh/config/sshd_config ]; then
+if [ ! -e "/mnt/data/ssh/config/sshd_config" ]; then
   cp -a /etc/ssh/sshd_config /mnt/data/ssh/config/sshd_config
 fi
-if [ ! -e /mnt/data/ssh/.ssh/authorized_keys ]; then
+if [ ! -e "/mnt/data/ssh/.ssh/authorized_keys" ]; then
   touch /mnt/data/ssh/.ssh/authorized_keys
 fi
 
 # Link to user config
-rm -Rf /etc/ssh
-ln -s /mnt/data/ssh/config /etc/ssh
-ln -s /mnt/data/ssh/.ssh ${USER_HOME}/.ssh
+if [ ! -L "/etc/ssh" ]; then
+  rm -Rf /etc/ssh
+  ln -s /mnt/data/ssh/config /etc/ssh
+fi
+if [ ! -L "${USER_HOME}/.ssh" ]; then
+  ln -s /mnt/data/ssh/.ssh ${USER_HOME}/.ssh
+fi
 
 # Initialize openssh-server
 mkdir -p /run/sshd
